Moved compatibility module to compat package.
